# backend-filemanagement/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My Cool API",
        "description": "An API of cool stuffs",
        "version": "1.0.0"
    },
    "paths": {
...
        "api/documents/store": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a single document",
                "description": "Uploads a single document and returns the uploaded document details",
                "operationId": "storeDocument",
                "requestBody": {
                    "description": "Document file to upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document uploaded successfully"
                    }
                }
            }
        },
        "api/documents/store-multiple": {
            "post": {
                "tags": [
                    "Documents"
                ],
                "summary": "Upload multiple documents",
                "description": "Uploads multiple documents and returns the uploaded documents details",
                "operationId": "storeMultipleDocuments",
                "requestBody": {
                    "description": "Array of document files to upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "files"
                                ],
                                "properties": {
                                    "files": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Documents uploaded successfully"
                    }
                }
            }
        },
        "api/documents": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Get paginated list of user's documents",
                "description": "Returns a paginated list of documents belonging to the authenticated user",
                "operationId": "getDocuments",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of documents per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "api/documents/find/{id}": {
            "get": {
                "tags": [
                    "Documents"
                ],
                "summary": "Find document by ID",
                "description": "Returns a single document by its ID if the user owns it",
                "operationId": "findDocumentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document found and returned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "403": {
                        "description": "Unauthorized to access the document"
                    }
                }
            }
        },
        "/documents/update/{id}": {
            "put": {
                "tags": [
                    "Documents"
                ],
                "summary": "Update document by ID",
                "description": "Updates an existing document by its ID if the user owns it",
                "operationId": "updateDocumentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "New document file",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Document updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Document updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "403": {
                        "description": "Unauthorized to update the document"
                    }
                }
            }
        },
        "/documents/destroy/{id}": {
            "delete": {
                "tags": [
                    "Documents"
                ],
                "summary": "Delete document by ID",
                "description": "Deletes an existing document by its ID if the user owns it",
                "operationId": "deleteDocumentById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the document to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Document deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Document deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found"
                    },
                    "403": {
                        "description": "Unauthorized to delete the document"
                    }
                }
            }
        },
...
    }
}